% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/df2phytaxa.R
\name{df2phytaxa}
\alias{df2phytaxa}
\title{Phylomatic format}
\usage{
df2phytaxa(taxa, uppercase = TRUE)
}
\arguments{
\item{taxa}{A data frame with columns named family, genus, and species.}

\item{uppercase}{logical. Should the function capitalize first letters?}
}
\description{
Phylomatic format
}

