% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixCase.R
\name{fixCase}
\alias{fixCase}
\title{Fix the name case of a taxon}
\usage{
fixCase(x)
}
\arguments{
\item{x}{a unit character vector with a taxon}
}
\description{
Fix the name case of a taxon
}
\examples{
fixCase("myrcia lingua")
fixCase("Myrcia Lingua")
fixCase("COFFEA ARABICA")
}

