% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flora.R, R/key.R
\docType{class}
\name{flora}
\alias{flora}
\alias{flora-class}
\alias{initialize,flora-method}
\alias{.allowableSystems}
\title{S4 definition for flora-class objects}
\usage{
\S4method{initialize}{flora}(.Object, ...)

.allowableSystems()

\S4method{initialize}{flora}(.Object, ...)

.allowableSystems()
}
\description{
flora-class objects store hierarchical key information for the floRa package

key-class objects connect flora-class objects to taxon-class objects
}
\details{
`system` can be one of "Cronquist", "APG1", "APG3", or "Unknown"

`key` must be a key-class object

`system` can be one of "Cronquist", "APG1", "APG3", or "Unknown"

`key` must be a key-class object
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of the flora}

\item{\code{key}}{Hierarchical key for the flora. See details.}

\item{\code{sytem}}{System used by the flora. See details.}

\item{\code{publisher}}{Publisher of the flora}

\item{\code{date}}{Date flora published}

\item{\code{ISBN}}{ISBN of the flora}

\item{\code{tree}}{data.tree object specifying}
}}

\author{
Brandon McNellis

Brandon McNellis
}
