% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.synonyms.R
\name{get.synonyms}
\alias{get.synonyms}
\title{List all synonyms of a given taxa}
\usage{
get.synonyms(taxon, fix = FALSE, relationship = FALSE)
}
\arguments{
\item{taxon}{a character vector containing a name.}

\item{fix}{should the function fix issues (synonyms, mispelled names) in taxon?}

\item{relationship}{return the kind of relationship?}
}
\value{
a character vector
}
\description{
This function returns all the synonyms of a given taxon according to the Brazilian Flora 2020.
}
\examples{
\dontrun{
get.synonyms("Myrcia lingua")
}
}
