% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_endemism.R
\name{get_endemism}
\alias{get_endemism}
\title{Get the listed endemism for a list of taxa}
\usage{
get_endemism(taxa)
}
\arguments{
\item{taxa}{A data frame with taxon names as returned by get.taxa()}
}
\value{
A data frame as returned by get.taxa with an extra column named vegtype.
}
\description{
Get the listed endemism for a list of taxa
}
\examples{
\dontrun{
taxa <- get.taxa(c("Myrcia guianensis", "bleh", "Xylopia", "Miconia albicans", "bleh", "Poa annua"))
get_endemism(taxa)
}
}
