% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lifeform.R
\name{get_lifeform}
\alias{get_lifeform}
\title{Get the listed vegetation life forms for a list of taxa}
\usage{
get_lifeform(taxa)
}
\arguments{
\item{taxa}{A data frame with taxon names as returned by get.taxa()}
}
\value{
A data frame as returned by get.taxa with an extra column named life.form
}
\description{
Get the listed vegetation life forms for a list of taxa
}
\examples{
\dontrun{
taxa <- get.taxa(c("Myrcia guianensis", "bleh", "Xylopia", "Miconia albicans", "bleh", "Poa annua"))
get_lifeform(taxa)
}
}
