% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plants.R
\docType{data}
\name{plants}
\alias{plants}
\title{Plant names}
\format{A character vector with 16 names}
\description{
A small character vector containing 16 plant names. Contains accepted names,
synonyms, and misspelled taxa.
}
\keyword{datasets}

