% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/remove.authors.R
\name{remove.authors}
\alias{remove.authors}
\title{Remove the author(s) from a taxon name.}
\usage{
remove.authors(taxon)
}
\arguments{
\item{taxon}{a character vector containing a single taxon}
}
\value{
a character vector
}
\description{
This attempts to remove the authorities of a taxonomic name.
}
\examples{
remove.authors("Coffea arabica L.")
remove.authors("Chrysophyllum argenteum subsp. nitidum (G.F.W.Meyer) T.D.Penn.")
}

