% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/standardize.names.R
\name{standardize.names}
\alias{standardize.names}
\title{Standardize taxonomic names}
\usage{
standardize.names(taxon)
}
\arguments{
\item{taxon}{a character vector containing a single name}
}
\value{
a character vector
}
\description{
This function standardizes taxa names. It is used mainly internally, but might be
helpful to the end user in some situations.
}
\examples{
standardize.names("Miconia sp 01")
standardize.names("Miconia Sp 2")
standardize.names("Sp18")
}

