% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon.R
\docType{class}
\name{taxon}
\alias{taxon}
\alias{taxon-class}
\alias{initialize,taxon-method}
\title{S4 definition for taxon-class objects}
\usage{
\S4method{initialize}{taxon}(.Object, ...)
}
\description{
taxon-class objects store taxonomic ranks for use in the floRa package
}
\details{

}
\section{Slots}{

\describe{
\item{\code{binomial}}{Binomial (Genus + Species) name}

\item{\code{variety}}{Variety}

\item{\code{species}}{Species}

\item{\code{subgenus}}{Subgenus}

\item{\code{genus}}{Genus}

\item{\code{tribe}}{Tribe}

\item{\code{subfamily}}{Subfamily}

\item{\code{family}}{Family}

\item{\code{order}}{Order}

\item{\code{division}}{Division}

\item{\code{citation}}{ICN Author citation}

\item{\code{taxon}}{Full taxonimic citation}
}}

\author{
Brandon McNellis
}
