% Generated by roxygen2 (4.0.1): do not edit by hand
\name{vernacular}
\alias{vernacular}
\title{Vernacular name search}
\usage{
vernacular(name, exact = FALSE)
}
\arguments{
\item{name}{a vernacular name}

\item{exact}{approximate or exact match?}
}
\value{
a data frame of results or NA
}
\description{
Search for taxa using vernacular names
}
\examples{
vernacular("pimenta", exact = TRUE)
vernacular("pimenta", exact = FALSE)
}

