% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2phytaxa.R
\name{df2phytaxa}
\alias{df2phytaxa}
\title{Phylomatic format}
\usage{
df2phytaxa(taxa, uppercase = TRUE)
}
\arguments{
\item{taxa}{A data frame with columns named family, genus, and species.}

\item{uppercase}{logical. Should the function capitalize first letters?}
}
\description{
Convert the results of get.taxa() to the phylomatic sample format
}
