% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.synonyms.R
\name{get.synonyms}
\alias{get.synonyms}
\title{List all synonyms of a given taxa}
\usage{
get.synonyms(taxon)
}
\arguments{
\item{taxon}{a character vector containing a name.}
}
\value{
a character vector
}
\description{
This function returns all the synonyms of a given taxon according to the Brazilian Flora 2020.
}
\examples{
\dontrun{
synonyms("Myrcia lingua")
}
}

