% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lower.taxa}
\alias{lower.taxa}
\title{Get downstream taxa}
\usage{
lower.taxa(taxon, accepted = TRUE)
}
\arguments{
\item{taxon}{a character vector with either a family or genus name}

\item{accepted}{list only accepted names?}
}
\description{
Get all downstream taxa from a family or genus name.
}
\examples{
lower.taxa("Acosmium")
lower.taxa("Zygophyllaceae")
}

