% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower.taxa.R
\name{lower.taxa}
\alias{lower.taxa}
\title{Get downstream taxa}
\usage{
lower.taxa(taxon, accepted = TRUE)
}
\arguments{
\item{taxon}{a character vector with either a family or genus name}

\item{accepted}{list only accepted names?}
}
\description{
Get all downstream taxa from a family or genus name.
}
\examples{
\dontrun{
lower.taxa("Acosmium")
lower.taxa("Zygophyllaceae")
}
}
